+++ nix/hosts/dionysus/default.nix
@@ -3,6 +3,7 @@
   imports = [
     ./hardware-configuration.nix
     ../../configuration.nix
+    ../../modules/steam.nix
   ];
 
   networking.hostName = "dionysus";
@@ -14,21 +15,5 @@
   hardware.graphics.extraPackages = with pkgs; [
     intel-media-driver
   ];
-
-  programs.steam = {
-    enable = true;
-    extest.enable = true;
-    package = pkgs.steam.override {
-      extraEnv = {
-        STEAM_FORCE_WAYLAND = "1";
-        SDL_VIDEODRIVER = "wayland";
-        GDK_BACKEND = "wayland,x11";
-        QT_QPA_PLATFORM = "wayland;xcb";
-      };
-    };
-    extraPackages = with pkgs; [ gamescope ];
-  };
-
-  programs.gamescope.enable = true;
 
 }